Bug fix: Fix hang
